Final french translation update
Now the French translation is fully finished and fully checked (by playing Ape and Wolf).
There will be no more updates, except for the ones you send me.
There are still some SMS and Kiwii messages that are not translated (I didn't have to translate them) and the choices are not translated either, since the last update of the game.